% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_map.R
\name{action_map}
\alias{action_map}
\title{Action Map}
\usage{
action_map(data, mun, cve_mpo, loc, week, num_loc, blocks)
}
\arguments{
\item{data}{it is the database of vector actions.}

\item{mun}{is the name of municipality.}

\item{cve_mpo}{is the id of municipality according the \href{**INEGI**}{https://www.inegi.org.mx/default.html}.}

\item{loc}{is the name of locality.}

\item{week}{is TRUE or FALSE. TRUE indicate the vector control actions in the week, else FALSE indicate the cumulative vector control.}

\item{num_loc}{is the id of locality according the \href{**INEGI**}{https://www.inegi.org.mx/default.html}.}

\item{blocks}{is the sf object of blocks.}
}
\value{
a tmap object.
}
\description{
This function generate the action map of vector control.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
