% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animap_areal.R
\name{animap_areal}
\alias{animap_areal}
\title{animap_areal}
\usage{
animap_areal(
  data,
  name,
  dir,
  country,
  breaks = NULLL,
  vel,
  save,
  cve_edo = NULL,
  dot = NULL,
  pal = NULL
)
}
\arguments{
\item{data}{is dengue dataset.}

\item{name}{is the name of the gif file.}

\item{dir}{is the directory where the animation will be saved.}

\item{country}{is logical value for indicating if the map is for all country o by state.}

\item{breaks}{is the numeric for defining the breaks of legend.}

\item{vel}{vel is the delay time between images. See also \link[tmap]{tmap_animation}.}

\item{save}{is the logical value for indicating if the gif is saved or not.}

\item{cve_edo}{is the id of state according the \href{**INEGI**}{https://www.inegi.org.mx/default.html}.}

\item{dot}{is the logical value, if dot TRUE the tmap is dot, else is areal.}

\item{pal}{is the palette.}
}
\value{
a gif file of animation.
}
\description{
Animated map areal of dengue cases by municipality
}
\details{
\link[tmap]{tmap_animation}.
}
\examples{
1+1
}
\seealso{
\link[tmap]{tmap_animation}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
