% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_cl.R
\name{cl_table}
\alias{cl_table}
\title{Generate a larval control table}
\usage{
cl_table(x, jur = NULL, mun)
}
\arguments{
\item{x}{is the dataset of control larvario.}

\item{jur}{is the Jurisdiccion.}

\item{mun}{is the municipio.}
}
\value{
a table.
}
\description{
Generate a larval control table
}
\details{
xxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[formattable]{formattable}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
