% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_heatmap.R
\name{den_heatmap}
\alias{den_heatmap}
\title{den_heatmap}
\usage{
den_heatmap(x, year, breaks)
}
\arguments{
\item{x}{is the dengue dataset.}

\item{year}{is the year dataset.}

\item{breaks}{is the breaks of legend.}
}
\value{
a ggplot object
}
\description{
heatmap of dengue cases by state.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
