% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendata_epichannel}
\alias{dendata_epichannel}
\title{Dengue cases by states for epidemiological channel.}
\format{
A df object with 1545 rows and 5 variables:
\describe{
  \item{DES_EDO_RES}{is the name of state.}
  \item{SEM}{Epidemiological week.}
  \item{q25}{percentil 25.}
  \item{q50}{percentil 50.}
  \item{q75}{percentil 75.}
  ...
}
}
\usage{
dendata_epichannel
}
\description{
is the dataset that contains the 25th, 50th and 75th percentiles of dengue cases by state.
}
\keyword{datasets}
