% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entomological_channel.R
\name{entomological_channel}
\alias{entomological_channel}
\title{Generate a Entomological Channel}
\usage{
entomological_channel(x, z, y, mun1, nom_loc, x_title, sep_ticks)
}
\arguments{
\item{x}{is the dataset of historic ovitraps. For more information, see \url{http://kin.insp.mx/aplicaciones/SisMV}.}

\item{z}{is the dataset of actual year of ovitraps.}

\item{y}{is the catalogue of localities accord the \href{**INEGI**}{https://www.inegi.org.mx/default.html}..}

\item{mun1}{is the name municipality.}

\item{nom_loc}{is the name of locality.}

\item{x_title}{is the name of tittle of x.}

\item{sep_ticks}{is the break in the x.}
}
\value{
A Entomological Channel.
}
\description{
This function has been designed to generate the
Entomological Channel of Aedes aegypti eggs.
}
\details{
The Entomological Channel is constructed in a similar way to the epidemiological channel, using the quantile function of the package stats. First we obtain the average number of eggs of the vector per year (a minimum time series of three years), week and location, second, the 25th, 50th and 75th percentiles are calculated, and categorized as success, safety, alert, respectively. As a third step, they are displayed with ggplot2 with the geom_area function. The fourth step includes the average number of eggs of the current year vector and is compared with the areas defined by the percentiles. If the temporary behavior of the eggs (average number per week) is below the success values (that is, under the area defined by success), then the control of egg abundance is in an area of success. This explanation is extrapolated to the security and alert areas. If the behavior of the eggs is above the warning zone, then it is in an epidemic zone, that is to say, the behavior of the eggs of the current year is superior to the behavior of the eggs of the time series.
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
