% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_channel.R
\name{epi_channel}
\alias{epi_channel}
\title{epi_channel}
\usage{
epi_channel(
  x,
  y,
  edo,
  jur,
  mpo,
  loc,
  year1,
  year2,
  x_epi,
  y_epi,
  x_alerta,
  y_alerta,
  x_seg,
  y_seg,
  x_exito,
  y_exito
)
}
\arguments{
\item{x}{is the database that contains the 25th, 50th and 75th percentiles of dengue cases by state}

\item{y}{is the current dataset. Load with \link[boldenr]{read_dataset_bol} function.}

\item{edo}{is the target state. The epidemiological channel is provided by state.}

\item{jur}{is the target furidiction.}

\item{mpo}{is the target municipality.}

\item{loc}{is the locality target.}

\item{year1}{is the year one.}

\item{year2}{is the year two.}

\item{x_epi}{is the x position of the epidemic zone legend.}

\item{y_epi}{is the y position of the epidemic zone legend.}

\item{x_alerta}{is the x position of the alert zone legend.}

\item{y_alerta}{is the y position of the alert zone legend.}

\item{x_seg}{is the x position of the security zone legend.}

\item{y_seg}{is the y position of the security zone legend.}

\item{x_exito}{is the x position of the success zone legend.}

\item{y_exito}{is the y position of the success zone legend.}
}
\value{

}
\description{
epi_channel
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
