% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiological_channel.R
\name{epidemiological_channel}
\alias{epidemiological_channel}
\title{Graph the epidemiological channel.}
\usage{
epidemiological_channel(x, edo, mun = NULL, state, scale_case, z, juris,
  jurisdiccion = NULL, year1, year2)
}
\arguments{
\item{x}{is the historic epidemiological dataset.}

\item{edo}{is the name state.}

\item{mun}{is the name municipality. is NULL for state level.}

\item{state}{is a logical parameter for graph the state or municipality. If the state is TRUE the graph is the state, else the Jurisdiction, municipality}

\item{scale_case}{is a numeric parameter for the breaks.}

\item{z}{is the current epidemiological dataset.}

\item{juris}{is a logical parameter for graph when the state is FALSE. if juris is FALSE the graph is the municipality}

\item{jurisdiccion}{is the name of jurisdiction. is null in the level municipality}
}
\value{
a graph of class ggplot.
}
\description{
This function has been designed to generate the
epidemiological channel of dengue cases.
}
\details{
The Entomological Channel is constructed in a similar way to the epidemiological channel, using the quantile function of the package stats. First we obtain the average number of eggs of the vector per year (a minimum time series of three years), week and location, second, the 25th, 50th and 75th percentiles are calculated, and categorized as success, safety, alert, respectively. As a third step, they are displayed with ggplot2 with the geom_area function. The fourth step includes the average number of eggs of the current year vector and is compared with the areas defined by the percentiles. If the temporary behavior of the eggs (average number per week) is below the success values (that is, under the area defined by success), then the control of egg abundance is in an area of success. This explanation is extrapolated to the security and alert areas. If the behavior of the eggs is above the warning zone, then it is in an epidemic zone, that is to say, the behavior of the eggs of the current year is superior to the behavior of the eggs of the time series.
}
\examples{
1+1

}
\seealso{
\link[boldenr]{entomological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
