% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_confirmados.R
\name{heatmap_confirmados}
\alias{heatmap_confirmados}
\title{Heatmap of dengue cases.}
\usage{
heatmap_confirmados(
  dataset,
  year,
  size_text,
  state,
  EDO = NULL,
  JS = NULL,
  MPO = NULL
)
}
\arguments{
\item{dataset}{is the dengue cases dataset.}

\item{year}{is the target year.}

\item{size_text}{is the font size of the text.}

\item{state}{is the target state.}

\item{EDO}{is a logical value, if EDO TRUE the heatmap is by state else by Jurisdiction or Municipality.}

\item{JS}{is a logical value, if JS TRUE the heatmap is by jurisdiction, else by municipality.}

\item{MPO}{is a logical value, if EDO TRUE the heatmap is by municipality, else the heatmap y by state_municipality.}
}
\value{

}
\description{
Heatmap of dengue cases.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
