% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_jur_hosp.R
\name{heatmap_jur_hosp}
\alias{heatmap_jur_hosp}
\title{Heatmap of hospitalized cases}
\usage{
heatmap_jur_hosp(
  x,
  state,
  year,
  rtitlesize,
  mtextsize,
  rlabelsize,
  clabelsize,
  rlabeltextsize,
  clabeltextsize,
  heatmap_color
)
}
\arguments{
\item{x}{is the sinave datasets. For more information, see \url{http://vectores.sinave.gob.mx/}.}

\item{state}{is the state.}

\item{year}{is the year.}

\item{rtitlesize}{is the row title size.}

\item{mtextsize}{is the matrix text size.}

\item{rlabelsize}{is the row label size.}

\item{clabelsize}{is the column label size.}

\item{rlabeltextsize}{is the row text label size.}

\item{clabeltextsize}{is the column text label size.}

\item{heatmap_color}{is character specifying the heatmap colour scheme.}
}
\value{
A heatmap of hospitalized cases by health jurisdiction.
}
\description{
Generates a Heatmap of hospitalized cases by health jurisdiction.
}
\details{
xxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
