% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_jur_prob_conf.R
\name{heatmap_jur_prob_conf}
\alias{heatmap_jur_prob_conf}
\title{Generate a heatmap of confirmed and probable cases by health jurisdiction}
\usage{
heatmap_jur_prob_conf(
  x,
  state,
  years,
  clus,
  prob,
  rtitlesize,
  mtextsize,
  rlabelsize,
  clabelsize,
  rlabeltextsize,
  clabeltextsize,
  heatmap_color
)
}
\arguments{
\item{x}{is the dataset of sinave. For more information, see \url{http://vectores.sinave.gob.mx/}}

\item{state}{is the state.}

\item{years}{is the current year.}

\item{clus}{is parameter for define the numbers of clusters.}

\item{prob}{is the status, 1 is probable, 2 positive casesa and 3 is negative cases.}

\item{rtitlesize}{is the row title size.}

\item{mtextsize}{is the matrix text size.}

\item{rlabelsize}{is the row label size.}

\item{clabelsize}{is the column label size.}

\item{rlabeltextsize}{is the row text label size.}

\item{clabeltextsize}{is the column text label size.}

\item{heatmap_color}{is character specifying the heatmap colour scheme.}
}
\value{
a heatmap of confirmed and probables cases by health jurisdiction.
}
\description{
Generate a heatmap of confirmed and probable cases by health jurisdiction
}
\details{
xxxxx
}
\examples{
1+1
}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
