% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_loc_prob_conf.R
\name{heatmap_loc_prob_conf}
\alias{heatmap_loc_prob_conf}
\title{Generate of heatmap of confirmed or probables cases by localities.}
\usage{
heatmap_loc_prob_conf(x, probable, year, state, rtitlesize, mtextsize,
  rlabelsize, clabelsize, rlabeltextsize, clabeltextsize, heatmap_color,
  n_loc)
}
\arguments{
\item{x}{is the dataset of sinave. For more information, see \url{http://vectores.sinave.gob.mx/}.}

\item{probable}{is the arguments for define the confirmed or probables cases. If probable is true the heatmap is the probables, else the heatmap is the confirmed cases.}

\item{year}{is the current year. Is a numeric.}

\item{state}{is the state.}

\item{rtitlesize}{is the row title size.}

\item{mtextsize}{is the matrix text size.}

\item{rlabelsize}{is the row label size.}

\item{clabelsize}{is the column label size.}

\item{rlabeltextsize}{is the row text label size.}

\item{clabeltextsize}{is the column text label size.}

\item{heatmap_color}{is character specifying the heatmap colour scheme.}

\item{n_loc}{is the number of localities.}
}
\value{
a heatmap confirmed or probables cases by localities.
}
\description{
Generate of heatmap of confirmed or probables cases by localities.
}
\details{
the heatmap is based in the package superheat and the function \link[superheat]{superheat}
}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}

\link[superhead]{superheat}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
