% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_jur.R
\name{hosp_jur}
\alias{hosp_jur}
\title{Heatmap of hospitalized dengue cases}
\usage{
hosp_jur(dataset, state, year, size_text)
}
\arguments{
\item{dataset}{is a dengue dataset.}

\item{state}{is the target state.}

\item{year}{is the year target}

\item{size_text}{is the font size of the text.}
}
\value{

}
\description{
Heatmap of hospitalized dengue cases
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
