% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_blocks.R
\name{hot_blocks}
\alias{hot_blocks}
\title{Generate a map of hot blocks by city.}
\usage{
hot_blocks(loc, x, blocks, cve_mpo, sem1, risk)
}
\arguments{
\item{loc}{is the locality.}

\item{x}{is the ovitraps dataset.}

\item{blocks}{is the the sf object of blocks the state.}

\item{cve_mpo}{is the id of municipality accord of INEGI.}

\item{sem1}{is the current week.}

\item{risk}{is the argument for map. If the risk is true the map is elaborate with the percentil and risk, else map the eggs by blocks.}
}
\value{
a map of risk or eggs map based in the percentil.
}
\description{
Generate a map of hot blocks by city.
}
\details{
xxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
