% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_conf_state.R
\name{map_conf_state}
\alias{map_conf_state}
\title{Generates a map of confirmed cases by municipalities}
\usage{
map_conf_state(x, y, cve_state, state, years, risk)
}
\arguments{
\item{x}{is the dataset of sinave.}

\item{y}{is sf object of all municipalities of  Mexico.}

\item{cve_state}{is the id of state, accord of INEGI.}

\item{state}{is the State.}

\item{years}{is the current year.}

\item{risk}{is the type map. If risk is true, the map is the risk, else the map is the cases.}
}
\value{
a map of risk or confirmed cases.
}
\description{
Generates a map of confirmed cases by municipalities
}
\details{
xxxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
