% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitraps_indicator.R
\name{ovitraps_indicator}
\alias{ovitraps_indicator}
\title{Generate a graph of ovitraps indicator,}
\usage{
ovitraps_indicator(x, nom_loc = NULL, all)
}
\arguments{
\item{x}{is the dataset of ovitraps of the current year.}

\item{nom_loc}{is the name of locality.}

\item{all}{is the argument for define the plot of all localities or one locality.}
}
\value{
a plot of ovitraps indicator.
}
\description{
Generate a graph of ovitraps indicator,
}
\details{
xxxx
}
\examples{
1+1

}
\references{


xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel} and \link[boldenr]{entomological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
