% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arbovirosis.R
\name{plot_arbovirosis}
\alias{plot_arbovirosis}
\title{Plot the probables and confirmed cases of arbovirosis of years 2018 and 2019 in Veracruz State}
\usage{
plot_arbovirosis(x, state, year1, year2)
}
\arguments{
\item{x}{is the dataset of probable and confirmed cases of arbovirosis.}

\item{state}{is the state target. its length is 1 o more.}

\item{year1}{is the previous year of the current year.}

\item{year2}{is the current year}
}
\value{
a plot the probables and confirmed cases of arbovirosis (DENV, CHIKV, ZIKV) of years 2018 and 2019 in Veracruz State. The lines and areas represent the cases of 2019 and 2018, respectively.
}
\description{
Plot the probables and confirmed cases of arbovirosis of years 2018 and 2019 in Veracruz State
}
\details{
xxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
