% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_serotype.R
\name{plot_state_serotype}
\alias{plot_state_serotype}
\title{Plot the dengue cases by state and serotype}
\usage{
plot_state_serotype(dataset, year)
}
\arguments{
\item{dataset}{is the dataset of sinave http://vectores.sinave.gob.mx/.}
}
\value{
a plot of class ggplot.
}
\description{
Plot the dengue cases by state and serotype
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
