% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_bol.R
\name{read_dataset_bol}
\alias{read_dataset_bol}
\title{Read the dataset of Sinave o vector}
\usage{
read_dataset_bol(path, dataset, inf = NULL)
}
\arguments{
\item{path}{is the directory where the files are located.}

\item{dataset}{is the parameter for define the type information.}

\item{inf}{is the parameter for define the type vector information. defaul is null for Sinave}
}
\value{
a dataset of sinave or vectors.
}
\description{
This function has been designed for read the dataset of sinave or vector
}
\examples{
1 + 1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
