% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_etv_data.R
\name{read_etv_data}
\alias{read_etv_data}
\title{read_etv_data}
\usage{
read_etv_data(x, etv, year, arbovirosis)
}
\arguments{
\item{x}{is the directory where the files are located.}

\item{etv}{is the parameter for define the etvs.}

\item{year}{is the year dataset.}

\item{arbovirosis}{is a logical value for define the group of etv, if is TRUE, the define DENV, CHIKV & ZIKV, else for other etvs.}
}
\value{
a dataframe
}
\description{
This function read the etvs txt file of sinave.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
