% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vbd.R
\name{read_vbd}
\alias{read_vbd}
\title{read_vbd}
\usage{
read_vbd(x, vbd, arbovirus)
}
\arguments{
\item{x}{is the directory where the files are located.}

\item{vbd}{is the parameter for define the vector-borne diseases.}

\item{arbovirus}{is a logical value for define the group of vector-borne diseases, if is TRUE, the define DENV, CHIKV & ZIKV, else for other etvs.}
}
\value{
a dataframe
}
\description{
This function read the vector-borne diseases txt file of sinave.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
