% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_percentil.R
\name{risk_percentil}
\alias{risk_percentil}
\title{Calculate the risk based in the percentils}
\usage{
risk_percentil(x, var, en)
}
\arguments{
\item{x}{is the dataset of class data.frame, tibble or data.table.}

\item{var}{is tha variable target for calculate the risk.}

\item{en}{is parameter for risk in english if en is true, else en is false the risk is in spanish.}
}
\value{
a object of same class of x.
}
\description{
Calculate the risk based in the percentils
}
\details{
xxxx
}
\examples{
1+1

}
\references{
xxxxx
}
\seealso{
\link[boldenr]{epidemiological_channel}, \link[stats]{quantile}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
