% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map_areal.R
\name{static_map_areal}
\alias{static_map_areal}
\title{static_map_areal}
\usage{
static_map_areal(
  x,
  breaks,
  week_start,
  week_end,
  country,
  cve_edo = NULL,
  year
)
}
\arguments{
\item{x}{is the dengue dataset.}

\item{breaks}{is the breaks of legend.}

\item{week_start}{is the week start for subset data.}

\item{week_end}{is the week end for subset data.}

\item{country}{is a logical parameter (TRUE or FALSE), if TRUE is all municipalities of Mexico else a state specific.}

\item{cve_edo}{is default NULL, but is id numeric for each state if country FALSE.}

\item{year}{is the year dataset.}
}
\value{
a ggplot object
}
\description{
static map of aggregated dengue cases by municipality
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
