% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_conf.R
\name{tb_conf}
\alias{tb_conf}
\title{Generates a table of the cases confirmed by institution or age groups}
\usage{
tb_conf(x, inst, year, state)
}
\arguments{
\item{x}{is the dataset.}

\item{inst}{is the parameter for define the type of table. If the inst is TRUE the table is by institution, else by age group.}

\item{year}{is the current year.}

\item{state}{is the state.

 @author Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}}
}
\value{
a formattable object.
}
\description{
Generates a table of the cases confirmed by institution or age groups
}
\examples{
1+1
}
