% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_ulv.R
\name{tb_ulv}
\alias{tb_ulv}
\title{Ultra Low Volumen (ULV) table.}
\usage{
tb_ulv(x, jur = NULL, mun, coldfog)
}
\arguments{
\item{x}{is the ULV datasets.}

\item{jur}{is the name of jurisdiction.}

\item{mun}{es valor logico TRUE o FALSE para indicar si la tabla es por jurisdiccion y municipia o por por estado y jurisdiccion, respectivamente.}

\item{coldfog}{is logical value TRUE or FALSE to indicate if the table is by jurisdiction and municipality or by state and jurisdiction, respectively}
}
\value{
a table.
}
\description{
This function generate of thermal o cold ULV table.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
