% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_epichannel.R
\name{den_epichannel}
\alias{den_epichannel}
\title{dengue epidemiological channel}
\usage{
den_epichannel(
  x,
  y,
  edo,
  jur,
  mpo,
  loc,
  current_year,
  last_year,
  x_epi,
  y_epi,
  x_alerta,
  y_alerta,
  x_seg,
  y_seg,
  x_exito,
  y_exito
)
}
\arguments{
\item{x}{is de dataset with dengue cases percentil.}

\item{y}{is the dengue dataset the current and last year.}

\item{edo}{is string indicating the state name.}

\item{jur}{is string indicating the jurisdiction name.}

\item{mpo}{is string indicating the municipality name.}

\item{loc}{is string indicating the locality name.}

\item{current_year}{is current year dengue dataset.}

\item{last_year}{is current year dengue dataset.}

\item{x_epi}{is the x position of the epidemic text.}

\item{y_epi}{is the y position of the epidemic text.}

\item{x_alerta}{is the x position of the alert text.}

\item{y_alerta}{is the y position of the alert text.}

\item{x_seg}{is the x position of the security text.}

\item{y_seg}{is the x position of the security text.}

\item{x_exito}{is the x position of the success text.}

\item{y_exito}{is the x position of the success text.}
}
\value{
a ggplot object.
}
\description{
den_epichannel generate the dengue epidemiological channel at level state, jurisdiction, municipality and locality.
}
\details{
if both jur and loc is NULL the function generate the dengue epidemiological channel of state, else only loc is NULL and jur provide the name of jurisdiction then generate the epidemiological channel at level jurisdiction.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
